Moved click_launcher to command_line directory, partially to avoid circular reference issues with Arch.
